Desktop System: Detect DeskTop already installed to RAMDisk properly
When a RAMCard has been identified, before copying there was a check
to see if DeskTop was already present. This was broken in many ways,
only one of which was my fault. The code did a lot of work to try
using the active volume and sniffing signature bytes from the slot...
all unnecessary.

Instead, just move up the code that computes the install path (e.g.
"/RAM5/DESKTOP") and check with a relative path. And then actually
handle the result properly if it is present.