Desktop System: Check size when detecting existin Desktop on RAMCard
This addresses a scenario reported on the socials that I've also seen
but didn't know how to repro. It reproduced in v1.2 and probably in
v1.1 as well.

When DESKTOP.SYSTEM identifies a RAMCard, and starts to copy files, it
does a CREATE on MODULES/DESKTOP then starts to do READ/WRITE pairs.
If the user hits Ctrl-OA-Reset before the CLOSE call (needed to set
the EOF), the empty file will be left on the RAMCard. Since it's a
very big file there's a long time window for this. (Note that Virtual
II clears RAM on a Ctrl-OA-Reset, so it's harder to repro there - you
need to Reset then boot your Omnidisk slot.)

When DESKTOP.SYSTEM starts the second time it sees the existing file
and skips the copy, trusting the RAMCard copy. When trying to launch
MODULES/DESKTOP the empty file is loaded at $2000 and invoked. But
that's where DESKTOP.SYSTEM itself was in memory, so it restarts in an
endless loop!

Some of the recent fixes change the observed behavior slightly from
v1.2 - it'll now basically hang with a black screen since it detects
the existing file and not try to re-copy.

Without verifying every existing file on the RAMCard, it's difficult
to prevent this entirely. But as a quick fix, ensure that the the
MODULES/DESKTOP file is more than a single block long.