Desktop System: Allow canceling copy to RAMCard with Escape
While the user can hold Apple or disable copying to RAMCard using the
Startup Options control panel, this requires advanced planning. And
planning is hard. We've learned (see previous commit) that users will
hit Ctrl+OA+Reset to restart.

Be nice and let users just hit Escape. A tip is shown.

This change shows a prompt during the "Copy DeskTop to RAMCard"
portion of the startup, and the user can hit Esc to abort the copy.
(Note that no cleanup is done, but the copy is treated as failed.)

If the user has shortcuts set up to copy apps to RAMCard "at boot"
that second sequence is unaffected, but can also be aborted with
Escape. The UI there needs an overhaul. And maybe the first Escape
should skip the second sequence entirely, but we can decide that
later.