DeskTop: Unify code and improve shortcuts/RAMCard behavior
There were 3 different implementations of code to start a shortcut -
one for invoking directly from the menu, another for invoking from the
dialog if not requiring a copy to RAMCard, and another for invoking
from the dialog if requiring a copy to RAMCard. Whee! Of course they
were all different with quirks.

Also, Desktop.system is nice and when it copies entries to the RAMCard
at boot it writes ENTRY_COPIED_FLAGS to the LCBANK2 stash. But DeskTop
never used these! Instead, DeskTop would look for the presence of a
file, which might be invalid if the copy had been canceled. So rework
DeskTop to check the flag, and also set the flag for "copy on use"
case.