Place module segments at block offsets
Modules (DeskTop, Selector, Disk Copy) use code "segments" - chunks of
code independently loaded, either at startup or as overlays. Modules
use DEFSEG up front to define the segment placement/length, which
define constants used by the loader code and within the segments
themselves to avoid repeating the data.

This change introduces BEGINSEG and ENDSEG macros for modules to use
around the segments. This replaces the leading .org and trailing
PAD_TO in most cases. And now that we have this we can (with tweaks to
the module loaders) introduce additional padding in the modules so
that the segments always occur on block boundaries. This should
improve loading performance somewhat.