DeskTop: Simplify "can we open this window?" icon calculations
DeskTop has a limit of 127 icons. So the number of additional icons we
can allocate when opening a window is the total number of icons minus
the number of icons actually allocated. (With slight complexity that
we reserve slots for all ProDOS devices plus Trash, so there's a bit
more math.)

But this used to be more complicated because icons weren't allocated
for list views! So instead the number of file records was used as a
proxy, which was really annoying and involved even more math since
there wasn't a simple count anywhere. But list views now use icons so
the code can be simplified.