International DA: Force full desktop repaint to avoid list view artifacts
If any DeskTop window was in list view, and the date format changed,
then later partial repaints could result in a mismatch in date format
between the newly painted areas and previously painted areas.

Force a full repaint when this setting changes - it's going to be an
infrequent setting so further optimization is unwarranted.