DeskTop: Update header only in active window to avoid artifacts
The "in disk" and "available" header numbers get updated for all
related windows (e.g. a volume window and folder window) when disk
operations are performed. But due to how repaints work, a partially
obscured header drawn before an update and then revealed after an
update could be drawd with different values, leading to visual
artifacts.

Introduce a cache of the values to draw for a window's header that is
updated from the true values only when the window is top-most,
refreshed when the window contents are refreshed (e.g. after a copy or
delete) or the window is activated.