BTK: Add shortcut support for radio/checkboxes
Rather than requiring shortcut keys (e.g Apple-1) to be given as part
of the label string for radio/checkbox controls, let them be specified
using the separate shortcut string, like for push buttons.

The only user-visible change is that in DeskTop's Add/Edit a Shortcut
dialog the shortcuts appear as suffixes rather than prefixes now.

The intent of this change is to standardize presentation of
radio/checkbox shortcuts (e.g. for other dialogs that don't have them
now) and allow more easily tweaking the presentation in the future.