DeskTop: Fix initializing default settings bug in loader
Regressed in 536a8074 - the "faux PRAM" segment was reduced to just
the minimal size - 256 bytes, with the current bell proc and settings.
It sits close to the top of memory but not at the very top, c/o
interrupt vectors. And it's no longer page-aligned. The loader assumed
all segment destination addresses were page-aligned. Oops.