DeskTop: Improve double-click responsiveness
Double-clicks can be missed if enough work needs to happen between the
mouse down and watching for the second click. This can happen if e.g.
multiple icons need to be deselected.

MGTK has a CheckEvents call that can be used to enqueue events during
periods between calls to GetEvent. DeskTop uses this during some
operations like I/O loops so the cursor isn't completely frozen.

Slip a call into the icon painting routine which are where most of the
time related to double-clicks is likely to go. In annecdotal tests
this seems to dramatically improve responsiveness.