DeskTop: Fix updating restart path when system path renamed
Regressed in 536a8074, but only because we got very very lucky before.
Maybe? The previous logic probably trashed something in Main LCBANK1
($D000-$DFFF) because a buffer placed there was used temporarily
during the path update, but LCBANK2 was banked in. Ooops. This is
called out in the test plan but really it was probably busted in v1.2,
and only noticed because the use of the LC banks was shuffled a bit in
the regression revision.

Use a free buffer elsewhere instead. Fixes it, yay.