DeskTop: Preserve selection across view changes
When changing views (e.g. from View > as Icons to View > by Name)
preserve the selection, if selection is in the window. This is done by
converting the selection list from icon ids (which are not preserved
across view changes) to record indexes (which are) before the old
icons are destroyed, then reversing the process once the new icons are
created.