Reset text window correctly when exiting/launching other programs
Noticed when invoking the BINSCII utility. Because the graphics mode
was still enabled when SETVID was called, that routine set the text
window for HGR mode with only four lines at the bottom of the screen.
Since BINSCII does not fully reset the display, its UI was cut off.

Flip the order so that SETVID is called after the display softswitches
are reset.