DeskTop: Replace operation dialogs with progress/alert dialogs
DeskTop uses a common "operation" concept to handle multi-stage
dialogs, including Get Info, Get Size, Copy/Move, "Download",
Lock/Unlock, and Delete. These use a state machine coupled to
infrastructure for processing the selection and recursively traversing
the files and directories. Complex, but it works.

One part that was a bit ugly is that a common dialog was used for all
of this. For example, when copying files, a single dialog box was
shown when the files were being enumerated with a count, while copying
to show the source, destination, and remaining count, but also to show
prompts and buttons if an issue was encountered (e.g. overwrite,
error, etc).

This meant a big ugly and mostly-empty dialog during the operation,
and then a non-standard prompt if unexpected anything happened.

Revise this to (1) use standard alerts if an issue comes up, including
if a Yes/No/All/Cancel prompt is needed, and (2) shrink the progress
dialog to a dedicated four-line dialog. The latter is positioned up
towards the top of the screen so that if an alert shows the relevant
filename is still visible.

This applies to:
* Copy/Move
* Lock/Unlock
* Delete
* Copy to RAMCard (a.k.a. "Download")

This saves about 380 bytes in main, but costs 200 in Aux and 130 in
LC. So still a net memory win.

In addition, a handful of strings (and l10ns) are improved. The
Yes/No/All buttons now show shortcuts if that option is enabled.