DeskTop: Keep mouse responsive during I/O operations
Since ProDOS MLI calls (which are mostly I/O bound) do bounce through
a relay already, slip in a call to MGTK's CheckEvents which allows the
mouse to move during e.g. file copies.

This might need tweaking if performance is impacted in a couple of
scenarios, e.g. GET_TIME or READ where the data is already buffered.
Doesn't seem to be a problem yet, though.