Disk Copy: Don't require LC segment to be padded to block boundary
Noticed in ebe3cccf with a quick fix, but introduced in fb23029d the
loader was updated to align segment boundaries with blocks to optimize
loading, but after the LC segment was loaded the mark was not set, so
the next incorrect was loaded incorrectly if padded to the next block.
Add a SET_MARK call and all is well.