Bootstrap: Reduce QUIT routine size and complexity
When invoked, both DeskTop and Selector use lib/bootstrap to install a
ProDOS QUIT routine. The routine resets the system, sets the right
PREFIX, and then loads the actual loader of the module to load all of
its segments.

The routine was a bit convoluted, first getting the PREFIX, patching
itself back into ProDOS, and invoking itself again.

With an eye towards reorganizing how memory is used in the ProDOS QUIT
routine area ($D100-$D3FF), streamline and simplify the routine,
chopping 98 bytes off of it.

No functional changes.