Disk Copy: Ensure a valid GrafPort is selected
DeskTop effectively hands MGTK state off to Disk Copy when the latter
is launched, so that the transition is clean. But the active port at
that point is no longer valid. Thus far we've been really, really
lucky that this doesn't cause problems, but further tinkering reveals
this is fragile.

Init and select a valid port before anything else is done with MGTK.