DeskTop: Fix path validation for volume icons
Root cause is a bug introduced in 146a3b321 but revealed by 1d9ceed2.
When composing icon paths (e.g. when opening an icon), for volume
icons there is no containing window path so a pointer is set to null.
But then this is incorrectly dereferenced to check the length.
Previously we got lucky, but now this at least initially points
somewhere causing the length check to fail. So yay?

There's no need to check the path length for volume icons, so just
skip that before doing the actual composition.