Loader: Ensure default settings/belldata are copied if needed
The change in 1d9ceed2 altered who was responsible for loading
settings. The logic worked fine of the files were present (load then
copy to RAM). But if the file load failed, the defaults were not
copied into the right place.

Rework the logic to ensure that something - either the defaults or the
loaded file - is always copied into place, for both settings and
sound.