Coding Style: Standardize on 'OK' instead of 'Ok'
'OK' and 'Ok' were used inconsistently in the code. 'OK' is what's
actually shown in dialogs and was slightly more common, so standardize
on that. This was particularly bad for alerts where the API was
inconsistent between button options and return values.

(Most uses of 'Ok' were from batch-conversions of labels like
'handle_ok_button' to 'HandleOkButton' rather than intentional.)