DeskTop: Enable Escape to cancel operations
The Delete and Copy/Move operations checked for the Escape key during
parts of the operation, but not during the initial count, and not when
iterating the selection. And other operations (Get Size, Lock, Unlock)
didn't check at all.

Rework to make the check code simpler and integrated into the
core operation logic.