Control Panel DA: Add button to set desktop pattern
Susan Kare's original, beautiful Macintosh control panel was designed
with no labels to simplify internationalization, and the interactivity
was fairly creative. To change the desktop pattern you'd doodle on the
fatbits display then click on the preview to apply it. Very compact
and simple, but not very discoverable.

If you grew up with that, the Control Panel DA works just the same
way. But I've seen a few people comment that how you apply the pattern
wasn't obvious. (e.g. http://toastytech.com/guis/adesk12.html)

So... turn the label into a button. You can still just click on the
preview, but this does improve discoverability and doesn't look too
bad.