This Apple DA: Fix crash showing memory count
The root cause is 1d9ceed2 but 68a40c29 is a culprit as well.
The DA calls into IntToStringWithSeparators to show the memory size.
That in turn calls into ReadSettings to get the separator. But...
This Apple is making the call from main and there is no ReadSettings
entry point defined. Is there? Well yes.... the definition in 68a40c29
just spills it into the current scope. And while many DAs defined
an 'aux' scope. This Apple didn't. So the symbol was exposed to
main bank code, called blithely, and boom.

A quick fix is to expose a new JUMP_TABLE_READ_SETTING API from
DeskTop for DAs, and alias ReadSettings to that. But to be more robust
and prevent this from happening, the DA_START_AUX_SEGMENT macro now
defines an 'aux' scope automatically and the aux entry point
definitions are scoped. This required changes to the source of
most DAs, but this commit only changes the binary for DeskTop (new
API) and the This Apple DA. The others build identically, and the
scoping and cross-bank calls are more obvious this way.