DeskTop: Support for launching AppleWorks files
c/o Hugh Hood's macro construction, a new AWLAUNCH.SYSTEM is included
in EXTRAS/ which acts as a proxy. The pathname passed in is inserted
into macro logic that opens the file.

This requires AppleWorks 5.1 to be installed to the same volume as
DeskTop, in a top-level directory named "AW5" e.g. /VOL/AW5. This is
where the installer would place it. Future changes may try to make
this more flexible.