AWLauncher.system: Update macros to fix loading pre-5.1 files
@frankmilliron noticed that AWP files were loading scrolled past the
bottom. Hugh Hood investigated and reported:

"If the WP document was originally saved with an earlier version of
AppleWorks (e.g. AppleWorks v2.1), AppleWorks 5.1 will load it and
display the cursor at or beyond the very end, UNLESS, the default load
setting under WP is changed to reset."

Hugh updated the macro as follows:

1. Is the file just loaded an AWP file?
2. If yes, does location $7C62/$7C63 contain $7D38?
3. If yes, issue an <OA-1> command to move the screen and cursor to
    the top of the file.

https://groups.google.com/g/comp.sys.apple2/c/lFRfyX0llhI/m/dZTLjNAJBAAJ

This moves the macro table to an external file that's slurped
into the build; we just need to keep the string offsets updated.