Image Preview: Fix HR image display glitches on palette bit transitions
In transitions between hires bytes with the palette bit set to not set,
the spill bit was set to 0. This would result in black DHR pixels being
introduced. Not very visible in most cases, but with black text on a
white background with palette switches within a line, glitches would
appear. This was noticed in the 'PIC.FILL.FAIL" image.