DeskTop: Improve positioning of window header items.
Improve the logic of laying out the header items in directory windows.
Previously, the logic pre-calculated the maximum space each chunk
(item count, k in disk, k available) used. Centering was wonky because
it assumed each item was full width, and the k available was not filly
right aligned either. It also had a minimum threshold below which the
layout logic shifted shifted entirely.

Change to computing the actual width of each chunk dynamically as it's
not in a tight loop. Then properly center the middle chunk between the
left and right chunks by computing the gap needed on each side. And
impose a lower limit on the gap. This makes the layout "perfect" for
all window sizes - cut off when it gets too low, but without a
transition.

Note that the middle chunk is not visually centered below the window
title, but it is visually centered with respect to the adjacent text
items. Can't satisfy both constraints!

Resolves #493