Disk Copy: Update block counts continuously during the copy
Disk Copy showed a "Reading..." / "Writing..." status as well as total
block count, blocks read count, and blocks written count. But the
status and counts were only updated when switching from reading to
writing mode or vice versa. This meant nothing to indicate the
progress of the copy other than the grinding of the disk.

Added code to update the counts as each block is read/written. This is
a little more CPU, but that's swamped by I/O. Much nicer.