DeskTop: Improve performance of Special > Get Size command
The Get Size command enumerates all selected icons and their
descendants. It would call GET_FILE_INFO on each to get the size of
the file in blocks to add to the total.

But since the enumeration walks over the directories, it implicitly
has the FileEntry structure for any descendants. And that already has
the block count, so the GET_FILE_INFO call is unnecessary!

Adjust the logic so that GFI is only called for the selected items
themselves, and use the FileEntry block count for descendants.

When doing Special > Get Size on a ProDOS master 5.25 floppy under
Virtual ][ at 1MHz, the time drops from 7 seconds to 4 seconds.