DeskTop: Fix file copy regression affecting Copy to RAMCard
Regression in e472ea84c - a necessary call to GET_FILE_INFO when
enumerating files was skipped, based on a misunderstanding of the
different enumeration phases. There's the initial "we're gonna tell
you the file name, you get the info" usually driven by enumerating the
selection, then there's recursion over directory contents. Only in the
latter case is the info pre-fetched for the proc.

The confusion about which proc does what is the motivation for a lot
of the recent cleanup.