DeskTop: Simplify dialog parameter blocks. Shaves 230 bytes.
Operation dialogs (About, Copy, Delete, New Folder, Get Info, Lock,
Rename, "Download", Get Size, and Duplicate) were invoked via an
indirection routine that took a dialog index and parameter block
pointer. The routine looked up the dialog proc and called it, passing
in the pointer. This is a potentially elegant approach that in theory
avoided global state and would allow two dialogs of the same type to
be displayed at the same time. But in practice this was unused and
meant lots of extra code to get at the parameters.

So just accept that this isn't as abstract as can be and (1) just call
the dialog procs directly and (2) have them access parameters
directly.

No functional changes.