DeskTop: Remove alert codes for single-use messages
Now that it's possible to use `ShowAlertParams` passing button options
and a string, there's no need to introduce an error code for every new
message. This saves 4 bytes in aux per message, at the expense of 4
bytes in main. But aux space is at more of a premium and it simplifies
the API and maintenance so... going for it.

L10n resource IDs for these and other alert strings are improved too.

With some other tweaks, delta is +31 bytes main, -48 bytes aux