Build: Use ca65 --warnings-as-error flag
Now that https://github.com/cc65/cc65/issues/1563 is fixed, replace
a hacky "anything on stderr fails the build" with the use of the
flag. Keep a colorizing script for stderr, though.

Note that this requires a VERY recent cc65 so you'll probably want to
update and build from source.