DeskTop: Fix "Files remaining" count after failed "copy to RAMCard"
When copying a shortcut to RAMCard, a hack was previously put in place
to ensure the count was correct if the shortcut target was at the root
of the disk - in that case, a directory is created corresponding to
the volume, which requires nudging the count by one.

But the hack set a flag that wasn't reset, so a subsequent copy of a
volume would end up with an off-by-one count.

The fix is to remove the special flag and just scope the hack to the
"copy to RAMCard" operation, which is now easy to distinguish.