DeskTop: Format/Erase - Show "Location" when prompting for name
This is mostly about filling a blank space in a big dialog, but serves
to remind the user which drive was selected. This is perhaps more
important if the command is activated with a volume icon selected, as
that will skip the device selection screen and go right to the name
prompt.