DeskTop: Rearrange overlay locations to simplify management
Previously, DeskTop had these overlays

* Format/Erase at $800 - $1A00
* File Dialog at $5000 - $7000
* Shortcut Picker at $9000 - $A000
* File Copy and Shortcut Edit at $7000 - $7800

This meant DeskTop needed to ensure that any routines used by overlays
were in $4000-$5000, or between $7800 and $9000, or after $A000.
Maintaining that intermediate hole was a pain as we've been steadily
reducing the code size.

The "good news everyone!" is that the file dialog is way, way smaller
now so we can crunch everything. The new scheme is:

* Format/Erase at $800 - $1A00 - unchanged
* Shortcut Picker at $5000 - $6000
* File Dialog at $6000 - $7000
* File Copy and Shortcut Edit at $7000 - $7800

The Shortcut Picker logic moves down so that the File Dialog routine
and the routines that drive the file dialog are adjacent. This is
important, as shortcut operations restore the latter in one go -
previously $5000-$7800, now $6000-$7800.

The only real complexity is that modal Desk Accessories were allowed
to use $5000-$7800 as a 10K buffer, so that's now added as a dedicated
restore chunk.

No functional changes.