DeskTop: Implement more efficient file move by relinking
When moving a file on the same volume, if block operations are
supported (i.e. not AppleTalk share), the move can be done without
copying the data through some subtle operations.

The approach taken in this case is:

* CREATE a file in the target location.
* Swap directory data between the source and destination file.
* DESTROY the file in the source location.

There's extra subtlety for directory moves as their key block has
references into the parent, so those are swapped too.

Resolves #350.