DeskTop: Default window view style to parent window's view style
When opening a window from an icon, use the parent window's style. For
volume icons, this will be "as Icons", but from folders in open
windows it will be the same as the parent window's style (e.g. "by
Name", "as Small Icons", etc)

Opening via path (e.g. Apple Menu, Shortcuts) will continue to default
to "as Icons".