DeskTop: Add "link" file support; add Make Link DA
This defines a new ProDOS file type - LNK ($E1/$0001) that acts as a
link (a.k.a. shortcut, a.k.a. alias) for another file path. It's
really quite minimal - just some header bytes and a ProDOS path. The
basic usage is... create a link, and you can open it / double-click it
as a shortcut to opening the original. Maybe you put links to games in
a directory. Or links to specific programs in the Apple menu.

Long term this might replace the current Shortcuts/Selector mechanism.
See #751 for thoughts there.

This commit has two parts:

* DeskTop get support for LNK files. An icon, type, and you can double
  click on them.

* A "Make Link" desk accessory that will create a link file for the
  selected icon (file or folder or volume). The links are created in
  a LINKS/ directory under the DeskTop folder, because they need to
  go somewhere.

The "Make Link" DA should be considered a temporary hack until I figure
out exactly what I want to do with Shortcuts/Selector.