MGTK: Allow keyboard control of menus even if initiated via mouse
If menu selection was initiated from the keyboard with Escape then
both the mouse and keyboard could be used to control menus, and
shortcuts (e.g. OA+Q) would be available.

If menu selection was initiated by the mouse, then only mouse control
was available and no keyboard control was available - no arrows, no
Return, not Escape, and no shortcuts.

This change makes MGTK implicitly invoke the menu-keyboard mode when
the menu is opened with the mouse too, which unfortunately required
more tinkering than expected because the "other" keyboard modes
(mousekeys and window moving/growing) shared the same code paths in
some cases.

Resolves #754