MGTK: Rework keyboard menu logic. Saves 151 bytes.
Keyboard menu support (rejiggered only slightly in 5582bf0e) was
basically a hack on top of mouse menu control. During event checking,
if keyboard control was enabled, then keys would effectively be
translated into mouse deltas, which would in turn drive the menu
selection. This caused the pointer to move during keyboard menu
operations, and to be nice necessitated position save/restore logic.

This is rewritten to make keyboard control inherent in the menu event
loop. The current highlighted item is now decoupled from the mouse
position, so keyboard control of the menu doesn't move the mouse at
all.

One feature or bug - I've heard it both ways - that goes away is that
the previously keyboard-selected menu item is not the first one shown
when Esc is pressed to enter the menu selection mode. Now, selection
always starts on the first menu. This may be less convenient for power
users, but is more predictable. Similarly, the unused and undocumented
SetMenuSelection MGTK call is removed, since it controlled this state.

As a side effect, this resolves #756, since menu behavior is not
dependent on having POSMOUSE work correctly.

And as usual in this code-base, any change which makes things work
more sensibly tends to save bytes.