DeskTop: Simplify window open/close animation code. Saves 95 bytes.
Simplify how the math is coded, use "asr16" macro and simple
twos-complement addition instead of flags and branching for negatives,
and helpers to get the icon bounds.

This changes slightly how the animation appears relative to the icon.
Previously, the animation's origin was a point near the top-left of
the icon. Now it's centered on the icon (including label) which looks
much better. Could be further improved, though.