DeskTop: Further improve window open/close animation. Saves 23 bytes.
Start the animation using the whole icon bounds, not just the center
point, and simplify the guts of algorithm too. Looks better and saves
bytes, of course. Unlike on classic MacOS the rect includes the icon
name, not just the bitmap.

The number of animation steps is also reduced from 11 to 7 because
with the rect starting off larger it's not necessary to shrink it as
many times to "look good".