DeskTop: Fix inclusion of last item in Apple Menu.
The "About This Apple II" menu item was introduced in 54ec8bb5 and
various constants relating to the size of the menu were updated... but
I missed one. Oops. This caused "Sort Directory" to not show up in the
default build. Fixed! Thanks to Jay Edwards for pointing this out.