DeskTop: Show open/close animation for accessories / file previews
Like windows that animate a frame from the icon to the window, do the
same for desk accessories. The logic is:

* If opening a DA directly from the icon - that's the source. If
* opening a DA from a menu, determine if the DA's containing window
* and icon are present. If so, animate from that icon. Otherwise, use
* the volume's icon. If previewing a file, animate the file icon, not
* the DA icon.

Also, rather than animating to a window frame, which isn't known until
after launch or might not exist, animate to/from the full desktop.
This matches what MacOS 7 does.