DeskTop: Show open/close animation for windows without folder icons
When opening/closing a window, a link to a "parent icon" is tracked
and used to dim the folder (and eventually un-dim it). This was also
used for the open/close animation.

In the cases where there wasn't a parent icon, including opening via
path, which occurs for folders in the Apple Menu, Shortcuts, and
restored windows, or when opened from an icon but then closing the
containing window, no animation was shown.

But now we have a handy dandy "find an icon to animate to" that just
uses the path, and will default to the volume. So simplify the code
and use that instead.

This ends up saving 8 bytes, yay!