Control Panel: Fix color previews of desktop pattern.
For color patterns, the offset of the bit pattern must be screen
relative. This means that the desktop pattern can't be simply used as
the fill pattern for the preview in the window, because pattern
rendering is offset by the window location. So the pattern is
"rotated" by some offset to align it with the desktop.

The dialog layout was tweaked in a44eb351d which changed the offset of
the preview and broke the color sync. This revealed that the offset
calculation was completely bogus. The real calculation ends up being
much simpler, and fixes this "for good".