IconTK: Remove IconEntry::id member. Saves 127 bytes in LC
Icons are typically referred to by ID, which is used in lookup tables
(one in DeskTop itself and one in IconTK - need to improve that) which
map ID to the IconEntry struct address. Previously, the struct
contained the ID as well, so if you only had a pointer you could find
the ID.

This was used in only a handful of places and those all had the ID
handy anyway so... remove it. That saves one byte per IconEntry which
is a big deal. Also saves a handful of bytes in Main/Aux c/o code
simplification. Also also, this revealed several places where the
IconEntry structure was assumed without using struct offsets or
asserts. So yay!

The biggest change is that the IconTK::AddIcon call now takes a proper
param block ({.byte id, .addr ptr}) rather than just a ptr.

No functional changes (hopefully!)