IconTK: Replace IconEntry::iconbits w/ type. Saves ~127 bytes.
Although it's very convenient to have each icon cart around a pointer
to its IconResource (bitmap/mask), that's a lot of duplicate pointer
data and we like saving memory.

Externalize DeskTop's IconType to IconResource mapping table and pass
it as an init param to IconTK. This simplifies initialization of
icons, and only slightly complicates the small number of places that
need an icon's bitmap/mask. This saves one byte per IconEntry, so 127
bytes. Because the table is moved from main to LC the actual
accounting is complicated - that costs LC memory but it's a net win.
Not that we're doing anything with it now, more headroom is nice.

For those keeping score, this means an IconEntry now just needs 16
bytes for the name, 4 for x/y coordinates, and 4 others including
type, state/flags, window, and record index.

"We can solve any problem by introducing an extra level of
indirection." -- David J. Wheeler