MGTK: ActivateCtl/UpdateThumb shouldn't redraw on no-op
Prevent scrollbar flashing in DeskTop when window contents are moved
but the view doesn't change. This could be done in DeskTop itself, but
avoiding the repaint from within MGTK is easy and a more comprehensive
fix.

This allows removing dedicated handling for this within the Listbox
lib, so saves code overall. Yay!