IconTK: Remove higlighted_list concept. Saves 131 bytes.
DeskTop maintains a list of selected icons. IconTK did too, although
the utility of it has dropped and icons had a state bit for it. The
main use was while dragging, to check properties of the dragged items
(e.g. Trash can be moved but not dropped), constructing the outline
polygons, and moving/repainting items at the end.

None of these are done in loops - just at the start and end of a drag
operation - so it's efficient enough to iterate over *all* icons and
skip the non-highlighted ones, given the memory savings.

To ensure we don't waste a bunch of bytes handling the
iteration/filtering in four places, a IterateHighlightedIcons proc is
added that takes a callback (in A,X). To make the code more readable,
the concept of a "lambda" (anonymous function) is introduced with
macros e.g.

  INVOKE_WITH_LAMBDA IterateHighlightedIcons
  sta icon_id
  ...
  rts
  END_OF_LAMBDA

Well, *I* think it's more readable.

No functional changes. Hopefully.