This Apple: Include IIgs banks E0/E1 in total memory count
The IIgs memory manager which detects the number of RAM banks in the
system checks only the "fast RAM" and doesn't check the "slow RAM" in
banks E0/E1. This means the count is off by 128K.

It also turns out that many emulators (e.g. KEGS before 1.19, and
current MAME) seem to get this wrong as well, and if you ask for 256K
they actually give you 256K of "fast RAM" and the real total is 384K.
Neato!