DeskTop: Fix crash when ejecting disk with multiple open windows.
When DeskTop detects that a disk has been ejected it builds a list of
all windows associated with that volume, then iterates over the list
and closes them.

Recent changes to improve window animation caused the list to be
overwritten, because the same "find windows..." logic was is used
there and the same output list is used. This caused a crash during
when the iteration repeated.

Simplify the logic to just repeat the list building and close the last
window, then build the list again etc. Imperceptably slower, but no
crash and less code.