DeskTop: Fix bogus duplicate disk error from Check All Drives
Regressed in e5981cd8, which didn't zero out cached_window_entry_list
entries when icons were deleted. This revealed a latent bug in the
code to check for duplicate icon names. The "is there already a volume
icon with this name?" CompareNames check looks at all entries in the
list, but at the point where it is called (1) the list size has been
incremented already but (2) the new icon number has not been emplaced.
When the list entries were zeroed, this would see an icon number of 0
and... well, remarkably not crash. When the list is not zeroed a
deleted icon number is still in the list, and so it would detect a
duplicate.

Just ensure the last (and not yet valid) list entry is not compared.