DeskTop: Don't paint volume icons over menu bar
Regressed in ab71e44f which stopped clamping icon drag polys to the
screen. A side effect is that volume icons can be placed under the
menu bar... but were sometimes rendered on top of it.

Fix is to adjust the relevant clipping rects when painting to exclude
the menu bar.