Text Preview: Fix updated when scrolling in long files
In very long files, scrolling by a single line (e.g. via arrow kwys)
doesn't move the scroll thumb. This means the scrollbar wouldn't
paint, which meant the window's GrafPort wouldn't be re-asserted - due
either to changes to not repaint the scrollbar on no-ops or not
preserving the zero page. Either way, this meant the active viewport
would not be updated, causing the display to use the old viewport and
scrolling appear to "stick" until the scroll thumb moved.

Fix by explicitly setting the port after noodling with it.