DeskTop: Fix crash when exactly 127 icons are present
Bug introduced when window entry tables were consolidated. When
storing a new window's entries, subsequent records need to be moved
up. The logic moved an entry before verifying that any needed to be
moved, a classic off-by-one, and this overflowed a table and overwrote
the start of another proc. Logic re-ordered to address this.