Disk Copy: Fix crash trying to play alert sounds
When BELLDATA was moved to the Main/LCBank2 this made hosting the Bell
procs in auxlc fail since the code would bank itself out. This was not
noticed for Disk Copy because alerts are typically silent.

Rejigger the placement, moving both the bell lib and it's dependent
speed control lib into main. Overall, the block buffer memory remains
the same so copies will be at the same speed as before.