Key Caps: Remove incorrect attempt to detect Platinum IIe
Not new in 1.3 but easy fix, and this was affecting all IIes and the
Laser 128 as well. The code was assuming that with the shift key mod
(standard on the Platinum IIe) that PB2 would read low by default, but
that's incorrect - PB2 reads high by default, and low when the shift
key is pressed. It also reads high if there is no mod, so it can't be
used to detect.