IconTK: Construct drag outlines even for hidden icons.
Code was in place to only construct drag outlines for visible icons.
This may have been needed when dragging was constrained to the screen
bounds, or perhaps it was an optimization. Either way, this makes it
unclear what is being dragged, so just remove this. A bit more
commentary in 4c4438a5.